requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
bert-score==0.3.13
bertviz==1.4.0
bitsandbytes==0.41.1
boto3==1.34.91
botocore==1.34.91
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
conda-pack==0.7.1
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.7
decorator==5.1.1
deepspeed==0.14.0
dill==0.3.7
distro==1.8.0
docstring_parser==0.16
einops==0.6.1
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
filelock==3.13.1
flash-attn==2.5.4
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
httpx-socks==0.9.1
huggingface-hub==0.17.3
idna==3.4
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mysql-connector-python==8.3.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
openai==1.30.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parso==0.8.4
peft==0.5.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.24.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-socks==2.4.4
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rich==13.7.1
rouge-score==0.1.2
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.10.1
selenium==4.21.0
sentencepiece==0.1.99
shtab==1.7.1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch @ file:///data/sunwangtao/torch-2.1.2%2Bcu121-cp38-cp38-linux_x86_64.whl#sha256=daa179bb558f78f2165db974a6744ec8de2ea71eb6aaf362bdae7616012c0302
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.35.0
trio==0.25.1
trio-websocket==0.11.1
triton==2.1.0
trl==0.8.0
typing_extensions==4.11.0
tyro==0.7.3
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.29.0
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
